+++ packages/typespec-azure-core/src/validate.ts
@@ -1,6 +1,7 @@
-import { Program } from "@typespec/compiler";
-import { checkEnsureVerb, checkRpcRoutes } from "./decorators.js";
+import type { Program } from "@typespec/compiler";
 import { checkPreviewVersion } from "./decorators/preview-version.js";
+import { checkEnsureVerb } from "./decorators/private/ensure-verb.js";
+import { checkRpcRoutes } from "./decorators/private/needs-route.js";
 
 export function $onValidate(program: Program) {
   checkRpcRoutes(program);