+++ packages/typespec-client-generator-core/src/http.ts
@@ -121,9 +121,9 @@ export function getSdkHttpOperation(
       });
     }
   }
-  const responsesWithBodies = [...responses.values(), ...exceptions.values()].filter((r) => r.type);
+  const successResponsesWithBodies = responses.filter((r) => r.type);
   const parameters = diagnostics.pipe(
-    getSdkHttpParameters(context, httpOperation, methodParameters, responsesWithBodies[0]),
+    getSdkHttpParameters(context, httpOperation, methodParameters, successResponsesWithBodies[0]),
   );
   filterOutUselessPathParameters(context, httpOperation, methodParameters);
   return diagnostics.wrap({